VG-1562: Fixes not working with SDC workspaces
Also, made a change so this task does the following only:

allows you to substitute an entire or partial string for a layer or
table's workspace path. It can't be used if the workspace type or
dataset name has changed. It is ideal for scenarios where drive letters
change, switch to UNC paths, update SDE connection file information, and
so on. See the Common Scenarios section below